requirements.txt
langchain
langchain-community
crewai
python-dotenv
crewai[tools]
gradio
agentops
tavily-python
youtube-search